+++ src/lib.rs
@@ -1,9 +1,10 @@
+// Trait for peaking Results
 pub trait Peak<T, E>
 where
     E: std::fmt::Debug,
 {
-    fn if_err(self, f: fn(&E) -> ()) -> Self;
-    fn if_ok(self, f: fn(&T) -> ()) -> Self;
+    fn peak_err(self, f: fn(&E) -> ()) -> Self;
+    fn peak_ok(self, f: fn(&T) -> ()) -> Self;
 }
 
 impl<T, E> Peak<T, E> for std::result::Result<T, E>
@@ -11,7 +12,7 @@ where
     E: std::fmt::Debug,
 {
     /// Runs a function on `std::result::Result` if it contains an Err variant.
-    fn if_err(self, f: fn(&E) -> ()) -> Self {
+    fn peak_err(self, f: fn(&E) -> ()) -> Self {
         if let Err(e) = &self {
             f(e);
         }
@@ -20,7 +21,7 @@ where
     }
 
     /// Runs a function on `std::result::Result` if it contains an Ok variant.
-    fn if_ok(self, f: fn(&T) -> ()) -> Self {
+    fn peak_ok(self, f: fn(&T) -> ()) -> Self {
         if let Ok(t) = &self {
             f(t);
         }
@@ -32,59 +33,43 @@ where
 #[cfg(test)]
 mod tests {
     use crate::Peak;
-    // =========================== if_ok() tests ============================
 
     #[test]
     #[should_panic]
-    fn test_if_ok_is_called() {
-        Ok::<(), ()>(()).if_ok(|_| panic!("Nice")).unwrap();
+    fn peak_ok_should_be_called_when_result_is_ok() {
+        Ok::<(), ()>(()).peak_ok(|_| panic!("Nice")).unwrap();
     }
 
     #[test]
     #[allow(unused_must_use)]
-    fn test_if_ok_should_not_be_called() {
-        Err::<(), ()>(()).if_ok(|_| panic!("if_ok should not be called"));
+    fn peak_ok_should_not_be_called_when_its_err() {
+        Err::<(), ()>(()).peak_ok(|_| panic!("peak_ok should not be called"));
     }
 
     #[test]
-    fn test_if_ok_uses_result_value() {
+    fn peak_ok_should_use_result_value() {
         Ok::<&str, ()>("OK")
-            .if_ok(|e| assert_eq!(e, &"OK"))
+            .peak_ok(|e| assert_eq!(e, &"OK"))
             .unwrap();
     }
 
     #[test]
     #[should_panic]
-    fn test_if_ok_uses_result_value_wrong_assert() {
-        Ok::<&str, ()>("OK").if_ok(|e| assert_eq!(e, &"")).unwrap();
-    }
-
-    // =========================== if_err() tests ===========================
-
-    #[test]
-    #[should_panic]
     #[allow(unused_must_use)]
-    fn test_if_err_is_called() {
-        Err::<(), ()>(()).if_err(|_| panic!("Nice"));
+    fn peak_err_is_called_when_result_is_err() {
+        Err::<(), ()>(()).peak_err(|_| panic!("Nice"));
     }
 
     #[test]
-    fn test_if_err_should_not_be_called() {
+    fn peak_err_should_not_be_called_on_ok_result() {
         Ok::<(), ()>(())
-            .if_err(|_| panic!("if_err should not be called"))
+            .peak_err(|_| panic!("peak_err should not be called"))
             .unwrap();
     }
 
     #[test]
     #[allow(unused_must_use)]
-    fn test_if_err_uses_result_value() {
-        Err::<(), &str>("error").if_err(|e| assert_eq!(e, &"error"));
-    }
-
-    #[test]
-    #[should_panic]
-    #[allow(unused_must_use)]
-    fn test_if_err_uses_result_value_wrong_assert() {
-        Err::<(), &str>("error").if_err(|e| assert_eq!(e, &""));
+    fn peak_err_should_use_result_value() {
+        Err::<(), &str>("error").peak_err(|e| assert_eq!(e, &"error"));
     }
 }