+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## 1.0.0 - 2022-10-31
+
+### Changed
+
+-   Modified method names from `if_err` and `if_ok` to `peak_err` and `peak_ok` respectively
+
+### Added
+
+-   Added tests for `std::result::Result<T, E>` support for `Peak`.
+
 ## 0.0.1 - 2022-10-30
 
 ### Added